requirements.txt
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
db-dtypes==1.0.0
google-api-core==2.7.1
google-auth==2.6.4
google-auth-oauthlib==0.5.1
google-cloud-bigquery==3.0.1
google-cloud-bigquery-storage==2.13.1
google-cloud-core==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
grpcio==1.44.0
grpcio-status==1.44.0
idna==3.3
numpy==1.22.3
oauthlib==3.2.0
packaging==21.3
pandas==1.4.2
pandas-gbq==0.17.4
proto-plus==1.20.3
protobuf==3.20.0
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydata-google-auth==1.4.0
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
urllib3==1.26.9
